Move tests under parsing module to access internals
